+++ cherrydan/Infrastructure/Manager/TabBarManager.swift
@@ -19,23 +19,3 @@ class TabBarManager: ObservableObject {
         }
     }
 }
-
-// MARK: - Custom View Modifier
-struct HideTabBarModifier: ViewModifier {
-    func body(content: Content) -> some View {
-        content
-            .onAppear {
-                TabBarManager.shared.hide()
-            }
-            .onDisappear {
-                TabBarManager.shared.show()
-            }
-    }
-}
-
-// MARK: - View Extension
-extension View {
-    func hideTabBar() -> some View {
-        self.modifier(HideTabBarModifier())
-    }
-}
